organize declarations by scope
