+++ src/beatmap/mapping/validator.ts
@@ -1,22 +1,22 @@
 import {
    DifficultySchema as V1DifficultySchema,
    InfoSchema as V1InfoSchema,
-} from '../schema/v1/declaration.ts';
+} from '../schema/v1/declaration/mod.ts';
 import {
    BPMInfoSchema as V2AudioSchema,
    DifficultySchema as V2DifficultySchema,
    InfoSchema as V2InfoSchema,
-} from '../schema/v2/declaration.ts';
+} from '../schema/v2/declaration/mod.ts';
 import {
    DifficultySchema as V3DifficultySchema,
    LightshowSchema as V3LightshowSchema,
-} from '../schema/v3/declaration.ts';
+} from '../schema/v3/declaration/mod.ts';
 import {
    AudioDataSchema as V4AudioSchema,
    DifficultySchema as V4DifficultySchema,
    InfoSchema as V4InfoSchema,
    LightshowSchema as V4LightshowSchema,
-} from '../schema/v4/declaration.ts';
+} from '../schema/v4/declaration/mod.ts';
 
 /** Data check version map for schema beatmap info. */
 export const infoCheckMap = {
