+++ src/beatmap/schema/external/declaration/mod.ts
@@ -1,0 +1,1 @@
+export * from './bookmarks.ts';
