+++ src/beatmap/schema/external/mod.ts
@@ -4,4 +4,4 @@
  * @module
  */
 
-export * from './declaration.ts';
+export * from './declaration/mod.ts';
