+++ src/beatmap/schema/mod.ts
@@ -1,4 +1,4 @@
-export * from './common/mod.ts';
+export * from './shared/mod.ts';
 
 export * as v1 from './v1/mod.ts';
 export * as v2 from './v2/mod.ts';
