+++ src/beatmap/schema/shared/declaration/mod.ts
@@ -1,0 +1,3 @@
+export * from './common.ts';
+
+export * from './custom/contributor.ts';
