+++ src/beatmap/schema/v1/mod.ts
@@ -5,13 +5,14 @@
  */
 
 export * from './compat/mod.ts';
+export * from './declaration/mod.ts';
 export * from './optimize/mod.ts';
+
 export * from './basicEvent.ts';
 export * from './bombNote.ts';
 export * from './bpmEvent.ts';
 export * from './colorBoostEvent.ts';
 export * from './colorNote.ts';
-export * from './declaration.ts';
 export * from './difficulty.ts';
 export * from './info.ts';
 export * from './infoBeatmap.ts';
