+++ src/beatmap/schema/v2/declaration/mod.ts
@@ -1,0 +1,8 @@
+export * from './audioData.ts';
+export * from './difficulty.ts';
+export * from './info.ts';
+
+export * from './custom/bookmark.ts';
+export * from './custom/bpmChange.ts';
+export * from './custom/colorScheme.ts';
+export * from './custom/editor.ts';
