+++ src/beatmap/schema/v2/mod.ts
@@ -4,6 +4,10 @@
  * @module
  */
 
+export * from './compat/mod.ts';
+export * from './declaration/mod.ts';
+export * from './optimize/mod.ts';
+
 export * from './arc.ts';
 export * from './audioData.ts';
 export * from './basicEvent.ts';
@@ -13,12 +17,8 @@ export * from './bombNote.ts';
 export * from './bpmEvent.ts';
 export * from './colorBoostEvent.ts';
 export * from './colorNote.ts';
-export * from './compat/mod.ts';
-export * from './custom.ts';
-export * from './declaration.ts';
 export * from './difficulty.ts';
 export * from './info.ts';
 export * from './obstacle.ts';
-export * from './optimize/mod.ts';
 export * from './rotationEvent.ts';
 export * from './waypoint.ts';
