+++ src/beatmap/schema/v3/declaration/custom/bookmark.ts
@@ -1,6 +1,6 @@
 import { number, object, optional, string, tuple, union } from '@valibot/valibot';
-import type { IBookmark, IBPMChange } from '../../../types/beatmap/v3/mod.ts';
-import type { InferObjectEntries } from '../helpers.ts';
+import type { IBookmark } from '../../../../../types/beatmap/v3/custom/bookmark.ts';
+import type { InferObjectEntries } from '../../../helpers.ts';
 
 /** Schema declaration for v3 custom `Bookmark` */
 export const BookmarkSchema = object<InferObjectEntries<IBookmark>>({
@@ -10,12 +10,4 @@ export const BookmarkSchema = object<InferObjectEntries<IBookmark>>({
       tuple([number(), number(), number()]),
       tuple([number(), number(), number(), number()]),
    ])),
-});
-
-/** Schema declaration for v3 custom `BPMChange` */
-export const BpmChangeSchema = object<InferObjectEntries<IBPMChange>>({
-   b: number(),
-   m: number(),
-   p: number(),
-   o: number(),
 });
