+++ src/beatmap/schema/v3/declaration/mod.ts
@@ -1,0 +1,5 @@
+export * from './difficulty.ts';
+export * from './lightshow.ts';
+
+export * from './custom/bookmark.ts';
+export * from './custom/bpmChange.ts';
