+++ src/beatmap/schema/v3/mod.ts
@@ -4,6 +4,10 @@
  * @module
  */
 
+export * from './compat/mod.ts';
+export * from './declaration/mod.ts';
+export * from './optimize/mod.ts';
+
 export * from './arc.ts';
 export * from './basicEvent.ts';
 export * from './basicEventTypesForKeywords.ts';
@@ -13,9 +17,6 @@ export * from './bpmEvent.ts';
 export * from './chain.ts';
 export * from './colorBoostEvent.ts';
 export * from './colorNote.ts';
-export * from './compat/mod.ts';
-export * from './custom.ts';
-export * from './declaration.ts';
 export * from './difficulty.ts';
 export * from './fxEventBox.ts';
 export * from './fxEventBoxGroup.ts';
@@ -33,6 +34,5 @@ export * from './lightTranslationEvent.ts';
 export * from './lightTranslationEventBox.ts';
 export * from './lightTranslationEventBoxGroup.ts';
 export * from './obstacle.ts';
-export * from './optimize/mod.ts';
 export * from './rotationEvent.ts';
 export * from './waypoint.ts';
