+++ src/beatmap/schema/v4/declaration/custom/info.ts
@@ -1,7 +1,7 @@
 import { object, optional, string } from '@valibot/valibot';
-import type { ICustomCharacteristic } from '../../../types/beatmap/v4/mod.ts';
-import { CharacteristicNameSchema } from '../common/declaration.ts';
-import type { InferObjectEntries } from '../helpers.ts';
+import type { ICustomCharacteristic } from '../../../../../types/beatmap/v4/custom/info.ts';
+import type { InferObjectEntries } from '../../../helpers.ts';
+import { CharacteristicNameSchema } from '../../../shared/declaration/mod.ts';
 
 export const CustomCharacteristicSchema = object<InferObjectEntries<ICustomCharacteristic>>({
    characteristic: CharacteristicNameSchema,
