+++ src/beatmap/schema/v4/declaration/mod.ts
@@ -1,0 +1,7 @@
+export * from './audioData.ts';
+export * from './common.ts';
+export * from './difficulty.ts';
+export * from './info.ts';
+export * from './lightshow.ts';
+
+export * from './custom/info.ts';
