+++ src/beatmap/schema/v4/mod.ts
@@ -4,6 +4,10 @@
  * @module
  */
 
+export * from './compat/mod.ts';
+export * from './declaration/mod.ts';
+export * from './optimize/mod.ts';
+
 export * from './arc.ts';
 export * from './audioData.ts';
 export * from './basicEvent.ts';
@@ -11,9 +15,6 @@ export * from './bombNote.ts';
 export * from './chain.ts';
 export * from './colorBoostEvent.ts';
 export * from './colorNote.ts';
-export * from './compat/mod.ts';
-export * from './custom.ts';
-export * from './declaration.ts';
 export * from './difficulty.ts';
 export * from './fxEventBox.ts';
 export * from './fxEventBoxGroup.ts';
@@ -32,6 +33,5 @@ export * from './lightTranslationEventBox.ts';
 export * from './lightTranslationEventBoxGroup.ts';
 export * from './njsEvent.ts';
 export * from './obstacle.ts';
-export * from './optimize/mod.ts';
 export * from './rotationEvent.ts';
 export * from './waypoint.ts';
